requirements.txt
django
git-python
natsort
bootstrap3
pygments
markdown